+++ zanata-war/src/main/java/org/zanata/service/impl/TranslationServiceImpl.java
@@ -35,9 +35,7 @@ import javax.annotation.Nonnull;
 import javax.enterprise.context.RequestScoped;
 import javax.persistence.EntityManager;
 
-import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Multimap;
 import lombok.extern.slf4j.Slf4j;
 
 import org.apache.commons.lang.StringUtils;
@@ -56,6 +54,7 @@ import org.zanata.dao.DocumentDAO;
 import org.zanata.dao.ProjectIterationDAO;
 import org.zanata.dao.TextFlowDAO;
 import org.zanata.dao.TextFlowTargetDAO;
+import org.zanata.events.DocStatsEvent;
 import org.zanata.events.DocumentLocaleKey;
 import org.zanata.events.DocumentUploadedEvent;
 import org.zanata.events.TextFlowTargetStateEvent;
@@ -95,6 +94,7 @@ import com.google.common.base.Optional;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 
 import static org.zanata.events.TextFlowTargetStateEvent.TextFlowTargetState;
 
@@ -148,6 +148,9 @@ public class TranslationServiceImpl implements TranslationService {
     @Inject
     private Event<TextFlowTargetStateEvent> textFlowTargetStateEvent;
 
+    @Inject
+    private Event<DocStatsEvent> docStatsEvent;
+
     @Transactional
     @Override
     public List<TranslationResult> translate(LocaleId localeId,
@@ -193,6 +196,7 @@ public class TranslationServiceImpl implements TranslationService {
                 projectIteration, hLocale);
 
         List<TextFlowTargetState> states = Lists.newArrayList();
+        Map<ContentState, Integer> contentStates = Maps.newHashMap();
 
         for (TransUnitUpdateRequest request : translationRequests) {
             HTextFlow hTextFlow =
@@ -249,9 +253,17 @@ public class TranslationServiceImpl implements TranslationService {
                         hTextFlowTarget.getVersionNum() == 0) {
                         HTextFlow textFlow = hTextFlowTarget.getTextFlow();
 
-                        states.add(new TextFlowTargetState(textFlow.getId(),
-                            hTextFlowTarget.getId(), hTextFlowTarget.getState(),
-                            currentState));
+                        TextFlowTargetState state =
+                                new TextFlowTargetState(textFlow.getId(),
+                                        hTextFlowTarget.getId(),
+                                        hTextFlowTarget.getState(),
+                                        currentState);
+                        states.add(state);
+                        contentStates = DocStatsEvent
+                            .updateContentState(contentStates,
+                                state.getNewState(),
+                                state.getPreviousState(),
+                                hTextFlow.getWordCount());
                     }
                     result.isSuccess = true;
                 } catch (HibernateException e) {
@@ -275,16 +287,25 @@ public class TranslationServiceImpl implements TranslationService {
             result.translatedTextFlowTarget = hTextFlowTarget;
             results.add(result);
         }
-        DocumentLocaleKey documentLocaleKey =
-            new DocumentLocaleKey(
-                sampleHTextFlow.getDocument().getId(), hLocale.getLocaleId());
+        if(!states.isEmpty()) {
+            DocumentLocaleKey documentLocaleKey =
+                new DocumentLocaleKey(
+                    sampleHTextFlow.getDocument().getId(), hLocale.getLocaleId());
 
-        TextFlowTargetStateEvent tftUpdatedEvent =
-            new TextFlowTargetStateEvent(documentLocaleKey,
-                projectIteration.getId(),
-                authenticatedAccount.getPerson().getId(),
-                ImmutableList.copyOf(states));
-        textFlowTargetStateEvent.fire(tftUpdatedEvent);
+            TextFlowTargetStateEvent tftUpdatedEvent =
+                new TextFlowTargetStateEvent(documentLocaleKey,
+                    projectIteration.getId(),
+                    authenticatedAccount.getPerson().getId(),
+                    ImmutableList.copyOf(states));
+            textFlowTargetStateEvent.fire(tftUpdatedEvent);
+
+            DocStatsEvent docEvent =
+                    new DocStatsEvent(documentLocaleKey, projectIteration.getId(),
+                            contentStates,
+                            Iterables.getLast(states).getTextFlowTargetId());
+
+            docStatsEvent.fire(docEvent);
+        }
         return results;
     }
 
@@ -779,7 +800,10 @@ public class TranslationServiceImpl implements TranslationService {
                 }));
         final int numPlurals = resourceUtils.getNumPlurals(document, locale);
 
-        List<TextFlowTargetState> states = Lists.newArrayList();
+        ImmutableList.Builder<TextFlowTargetState> states =
+                ImmutableList.builder();
+
+        Map<ContentState, Integer> contentStates = Maps.newHashMap();
 
         for (TextFlowTarget incomingTarget : batch) {
             String resId = incomingTarget.getResId();
@@ -870,28 +894,43 @@ public class TranslationServiceImpl implements TranslationService {
                     hTarget.setCopiedEntityId(null);
                     textFlowTargetDAO.makePersistent(hTarget);
 
-                    states.add(new TextFlowTargetState(textFlow.getId(),
-                        hTarget.getId(), hTarget.getState(),
-                        currentState));
+                    TextFlowTargetState state =
+                        new TextFlowTargetState(textFlow.getId(),
+                            hTarget.getId(), hTarget.getState(), currentState);
+
+                    states.add(state);
+
+                    contentStates = DocStatsEvent
+                        .updateContentState(contentStates, state.getNewState(),
+                            state.getPreviousState(), textFlow.getWordCount());
                 }
             }
             if (handleOp.isPresent()) {
                 handleOp.get().increaseProgress(1);
             }
         }
-        Long actorId =
-            assignCreditToUploader ? authenticatedAccount.getPerson().getId() :
-                null;
+        ImmutableList<TextFlowTargetState> list = states.build();
+        if (!list.isEmpty()) {
+            Long actorId =
+                assignCreditToUploader ? authenticatedAccount.getPerson().getId() :
+                    null;
 
-        DocumentLocaleKey documentLocaleKey =
-            new DocumentLocaleKey(
-                document.getId(), locale.getLocaleId());
+            DocumentLocaleKey documentLocaleKey =
+                new DocumentLocaleKey(
+                    document.getId(), locale.getLocaleId());
 
-        TextFlowTargetStateEvent tftUpdatedEvent =
-            new TextFlowTargetStateEvent(documentLocaleKey,
-                projectIterationId, actorId, ImmutableList.copyOf(states));
-        textFlowTargetStateEvent.fire(tftUpdatedEvent);
+            TextFlowTargetStateEvent tftUpdatedEvent =
+                    new TextFlowTargetStateEvent(documentLocaleKey,
+                            projectIterationId, actorId, list);
+            textFlowTargetStateEvent.fire(tftUpdatedEvent);
 
+            DocStatsEvent docEvent =
+                new DocStatsEvent(documentLocaleKey, projectIterationId,
+                    contentStates,
+                    Iterables.getLast(tftUpdatedEvent.getStates()).getTextFlowTargetId());
+
+            docStatsEvent.fire(docEvent);
+        }
         textFlowTargetDAO.flush();
         return changed;
     }